% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CompareDb.R
\name{CompareDb}
\alias{CompareDb}
\title{Compare two data frames.  If input is matrix, will try to convert matrix
into data fram (via as.data.frame) to compare.}
\usage{
CompareDb(db.old, db.new, id.var)
}
\arguments{
\item{db.old}{A data frame.}

\item{db.new}{A data frame.}
}
\value{
Prints out a report on the difference between the db's to the console
}
\description{
Compare two data frames.  If input is matrix, will try to convert matrix
into data fram (via as.data.frame) to compare.
}
\examples{
a <- cbind("id"=c(1:5),"age"=c(50,60,20,87,41),"grade"=c(1,2,3,1,2))
b <- cbind("id"=c(1:5),"age"=c(50,60,20,87,41),"grade"=c(1,2,3,1,2))
c <- cbind("id"=c(1:5),"age"=c(50,60,20,87,40),"grade"=c(1,2,3,2,2))
d <- as.data.frame(cbind(c,"nodestat"=c(0,1,1,0,0)))
CompareDb(a,b,"id")
CompareDb(a,c,"id")
CompareDb(a,d,"id")
}

