% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isParsableToNumeric.R
\name{isParsableToNumeric}
\alias{isParsableToNumeric}
\title{Is the string parsable to a number?}
\usage{
isParsableToNumeric(x)
}
\arguments{
\item{x}{character string}
}
\value{
logical. Returns \code{TRUE} if \code{x} is parsable to a numeric (i.e.
\code{x} is either already in numeric format or is a number formatted
as a character)
}
\description{
Checks whether a string can be parsed into a number. Blanks are ignored.
}
\examples{
isParsableToNumeric("example")
isParsableToNumeric("123")
isParsableToNumeric(123)
isParsableToNumeric(NA)
}
\author{
Samuel Leung
}

