% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NanoStringQC.R
\name{NanoStringQC}
\alias{NanoStringQC}
\title{QC for NanoString Data}
\usage{
NanoStringQC(raw, exp, detect = 80, sn = 150, plots = TRUE, ttl = " ",
  explore = TRUE)
}
\arguments{
\item{raw:}{matrix of raw counts obtained from nCounter (rows are genes). The first three columns must be labeled: c("Code.Class","Name","Accession") and contain that information.}

\item{exp:}{matrix of annotations with rows in the same order as the columns of raw.}

\item{detect:}{Percent threshold of genes over lod that we would like to detect (not decimal).}

\item{sn:}{Signal to noise ratio of the housekeeping genes we are willing to tolerate}

\item{plots:}{Binary, indicates whether plots to visualise the results are requested, defaults to TRUE}

\item{ttl:}{a string to show the title on the plots}

\item{explore:}{returns the plots only, defaults to TRUE}
}
\value{
matrix of annotations updated with normalization parameters
}
\description{
This function computes NanoString quality control parameters and flags
}

