% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CombineOVCS.R
\name{CombineOVCS}
\alias{CombineOVCS}
\title{Combines Ovarian Codeset 1 and Codeset 2 Nanostring Data}
\usage{
CombineOVCS(CS1, CS2)
}
\arguments{
\item{CS1}{matrix of raw counts obtained from nCounter (rows are genes) with codeset 1.}

\item{CS2}{matrix of raw counts obtained from nCounter (rows are genes) with codeset 2.}
}
\value{
A matrix of combined data with the first three columns labeled as c("Code.Class", "Name", "Accession")
}
\description{
Matches genes that are common to Codeset 1 and Codeset 2 (same accession number) and combines them after fixing nomenclature.
}
\details{
The four genes "DUSP4", "IGKC", "LDHA", and "PTHLH" have non-matching accession and are added to both codesets. The first three columns
of \code{CS1}, and \code{CS2} must be labeled \code{c("Code.Class", "Name", "Accession")} and contain that information.
}
\examples{
library(otta)
data(rawOVCA2, rawPROT)
cs1 <- rawOVCA2; cs2 <- rawPROT
CombineOVCS(cs1[, 1:3], cs2[, 1:3])
CombineOVCS(cs1, cs2)
}
\author{
Aline Talhouk, Derek Chiu
}

