% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CombineOVCS.R
\name{CombineOVCS}
\alias{CombineOVCS}
\title{Combines Ovarian Codeset 1 and Codeset 2 Nanostring Data}
\usage{
CombineOVCS(CS1, CS2)
}
\arguments{
\item{CS1}{matrix of raw counts obtained from nCounter (rows are genes) with codeset 1. The first three columns must be labeled: c("Code.Class", "Name", "Accession") and contain that information.}

\item{CS2}{matrix of raw counts obtained from nCounter (rows are genes) with codeset 2. The first three columns must be labeled: c("Code.Class", "Name", "Accession") and contain that information.}
}
\value{
A matrix of combined data with the first three columns labeled as c("Code.Class", "Name", "Accession")
}
\description{
This function matches the genes that are common to Codeset 1 and Codeset 2 (same accession number) and combines them after fixing nomenclature.
}
\examples{
library(otta)
data(rawOVCA2, rawPROT)
cs1 <- rawOVCA2; cs2 <- rawPROT
CombineOVCS(cs1[, 1:3], cs2[, 1:3])
}
\author{
Aline Talhouk
}

